Moved AbstractContextualStorageManager to internal package after making public
